requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
sqlalchemy==2.0.25
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
boto3==1.34.14
moto==4.2.12
pyyaml==6.0.1
pandas==2.1.4
numpy==1.26.3
spacy==3.7.2
transformers==4.36.2
torch==2.1.2
torchvision==0.16.2
beautifulsoup4==4.12.2
scrapfly-sdk==0.8.19
google-cloud-bigquery==3.14.1
google-cloud-vision==3.5.0
pillow==10.2.0
scipy==1.11.4 
supertokens-python==0.2.1